+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 
+* Trigger change event on input target when date picked.
 * Fix adding of keydown handler to non-button toggle target.